ai_service/requirements.txt
fastapi
uvicorn
ultralytics
opencv-python-headless
numpy
torch
Pillow
python-multipart
google-generativeai
python-dotenv
